phase2/backend/requirements.txt
fastapi==0.109.0
sqlmodel==0.0.14
psycopg[binary]>=3.1.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
pydantic-settings>=2.7.0
uvicorn[standard]==0.27.0
